moving layoutBase to component folder
