+++ src/Components/Layout/LayoutBase.js
@@ -4,34 +4,11 @@ import Container from '@material-ui/core/Container';
 import PropTypes from 'prop-types';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import clsx from 'clsx';
-import TopBar from '../../Components/TopBar/TopBar';
-import Menu from '../../Components/Menu/Menu';
-import configs from '../../settings';
+import TopBar from '../TopBar/TopBar';
+import Menu from '../Menu/Menu';
+import styles from './_styles';
 
-
-const useStyles = makeStyles((theme) => ({
-    root: {
-        display: 'flex',
-        flexGrow: 1,
-    },
-    content: {
-        display: 'flex',
-        width: '100%',
-    },
-    containerBase: {
-        transition: theme.transitions.create(['width'], {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-        width: '100%',
-        marginTop: 70,
-        display: 'flex',
-    },
-    containerBaseShift: {
-        width: `calc(100% - ${configs.drawerWidth}px)`,
-    },
-
-}));
+const useStyles = makeStyles(styles);
 
 
 const LayoutBase = (props) => {
